Limit maximum supported resolution

Closes #10

See merge request videolan/x264!14

+++ common/base.h
@@ -263,7 +263,7 @@ X264_API void x264_log_internal( int i_level, const char *psz_fmt, ... );
 
 /* x264_malloc : will do or emulate a memalign
  * you have to use x264_free for buffers allocated with x264_malloc */
-X264_API void *x264_malloc( int );
+X264_API void *x264_malloc( int64_t );
 X264_API void  x264_free( void * );
 
 /* x264_slurp_file: malloc space for the whole file and read it */
@@ -296,7 +296,7 @@ do {\
 
 #define PREALLOC_INIT\
     int    prealloc_idx = 0;\
-    size_t prealloc_size = 0;\
+    int64_t prealloc_size = 0;\
     uint8_t **preallocs[PREALLOC_BUF_SIZE];
 
 #define PREALLOC( var, size )\
